Adding the ability to cross-match with G4Jy survey

+++ scripts/run_pipeline.sh
@@ -16,15 +16,18 @@ else
 	DB_update="${DB_dir}db_update_log.py"
 	DB_end="${DB_dir}db_end_log.py"
 
-	# Directory where data needs to be stored at the end of the pipeline
+	# Change to directory where data needs to be stored at the end of the pipeline
 	data_dir="/data/awaszewski/project/obsids_109/${obsid}/"
-	# Directory on Garrawarla where second half of pipeline will run
+	# Change to directory on Garrawarla where second half of pipeline will run
 	garra_dir="/astro/mwasci/awaszewski/EoR_scin_pipeline/"
+	# Change to directory where pipeline is being controlled from
+	pipeline_dir="/data/awaszewski/EoR_scin_pipeline/"
 
 	# Set all pipeline commands
-	scripts="/data/awaszewski/EoR_scin_pipeline/scripts/"
-	moment="${scripts}moment_image.sh"
-	vot="${scripts}make_vot.sh"
+	scripts="${pipeline_dir}/scripts/"
+	moment="${scripts}/moment_image.sh"
+	vot="${scripts}/make_vot.sh"
+	match="${scripts}/match_g4jy.sh"
 
 	# Add current observation to database
 	python3 $DB_add -o ${obsid} -d ${data_dir}
@@ -57,18 +60,25 @@ else
 					running=1
 				fi
 			done
-			
-			# Final clean up of /astro, including measurement set
-			scp hpc-data:${garra_dir}currently_running/${obsid}* $data_dir
-			ssh garrawarla "cd ${garra_dir}currently_running/; rm ${obsid}*; cd /astro/mwasci/asvo/${asvojob}/; rm ${obsid}.ms"
 		
-			# Moment images
-			bash $moment ${obsid}
+			if [[ "${output}" == *"Failed"* ]]; then
+				echo ${obsid} ${index} >> processing_failed.txt
+			else	
+				# Final clean up of /astro, including measurement set
+				scp hpc-data:${garra_dir}currently_running/${obsid}* $data_dir
+				ssh garrawarla "cd ${garra_dir}currently_running/; rm ${obsid}*; cd /astro/mwasci/asvo/${asvojob}/; rm ${obsid}.ms"
+			
+				# Moment images
+				bash $moment ${obsid}
 
-			# Source finding
-			bash $vot ${obsid}
+				# Source finding
+				bash $vot ${obsid}
 
-			echo ${obsid} ${index} >> processing_complete.txt
+				# Cross-match with G4Jy survey
+				bash $match ${obsid}
+
+				echo ${obsid} ${index} >> processing_complete.txt
+			fi
 		else
 			# If communication with garrawarla fails pipeline will be terminated
 			python3 $DB_end -o ${obsid} -s Failed
